+++ cmd/cortextool/pkg/rules/rules_test.go
@@ -138,11 +138,18 @@ func TestLintPromQLExpressions(t *testing.T) {
 			err: "",
 		},
 		{
+			name:     "with a complex expression",
+			expr:     `sum by(cluster, namespace) (sum_over_time((rate(loki_distributor_bytes_received_total{job=~".*/distributor"}[1m]) * 60)[1h:1m])) / 1e+09 / 5 * 1 > (sum by(cluster, namespace) (memcached_limit_bytes{job=~".+/memcached"}) / 1e+09)`,
+			expected: `sum by(cluster, namespace) (sum_over_time((rate(loki_distributor_bytes_received_total{job=~".*/distributor"}[1m]) * 60)[1h:1m])) / 1e+09 / 5 * 1 > (sum by(cluster, namespace) (memcached_limit_bytes{job=~".+/memcached"}) / 1e+09)`,
+			count:    1, modified: 0,
+			err: "",
+		},
+		{
 			name:     "with an invalid expression",
 			expr:     "it fails",
 			expected: "it fails",
 			count:    0, modified: 0,
-			err: "parse error at char 4: could not parse remaining input \"fails\"...",
+			err: "1:4: parse error: unexpected identifier \"fails\"",
 		},
 	}